Plow.Extras.List

Documentation

groupUp :: Int -> [t] -> [[t]]

groupUp is a function that takes a list and turns it into a list of n sized lists. except the last list which is whatever size there is left very useful for concurrentIO on list